perf(tests): Reduce TTL/timeout test sleeps from seconds to milliseconds
Problem: 5 tests used 1-2 second sleep durations when testing TTL/timeout
behavior, wasting ~7 seconds of test runtime.

Changes:
- Reduce stuck_threshold_seconds from 2.0/1.0 to 0.05 (50ms)
- Reduce ttl_seconds from 1 to 0.05 (50ms)
- Reduce time.sleep from 2.0/1.5/1.1 to 0.1 (100ms)
- Change pubsub.py ttl_seconds type from int to float for sub-second precision

Tests affected:
- test_scenario_verifier_detects_phase_timeout
- test_scenario_stuck_phase_severity_is_warning
- test_scenario_expired_messages_move_to_dead_letter
- test_scenario_dead_letter_messages_can_be_retried
- test_findings_expire_after_ttl

Result: ~7.3 seconds saved per test run

+++ cortical/reasoning/pubsub.py
@@ -70,7 +70,7 @@ class Message:
     payload: Dict[str, Any]
     sender: str
     timestamp: datetime
-    ttl_seconds: Optional[int] = None
+    ttl_seconds: Optional[float] = None  # float for sub-second precision in tests
     priority: int = 0
     retry_count: int = 0
     status: MessageStatus = MessageStatus.PENDING
@@ -240,7 +240,7 @@ class PubSubBroker:
         topic: str,
         payload: Dict[str, Any],
         sender: str,
-        ttl_seconds: Optional[int] = None,
+        ttl_seconds: Optional[float] = None,
         priority: int = 0,
     ) -> str:
         """
@@ -550,7 +550,7 @@ class PubSubBroker:
 
         return sorted(messages, key=lambda m: m.timestamp, reverse=True)
 
-    def retry_dead_letter(self, message_id: str, new_ttl_seconds: Optional[int] = 300) -> bool:
+    def retry_dead_letter(self, message_id: str, new_ttl_seconds: Optional[float] = 300) -> bool:
         """
         Retry delivery of a dead letter message.
